backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
alembic==1.13.1
python-jose[cryptography]==3.3.0
gunicorn==21.2.0
bcrypt==4.0.1
email-validator>=2.0.0